Made Message sorted collection of fields instead of RandomAccessSeq.
Message is just a sorted collection of fields with no random access as it
seems to be not needed and could limit optimization techniques.
Moreover Message should be immutable structure and if one needs mutable
version should use MessageBuilder instead.

MessageBuilder does not inherit from Message as this class is not really
a Message. This is similar to situation with String and StringBuilder.

Signed-off-by: Grzegorz Kossakowski <grek@google.com>