Method children in UserType should return Seq of NestedMember[_].
NestedMember[_] which is short-hand for existential type must be used
because NestedMember is nonvariant type which seems to make sense.

Intuitively, existential type represents fact that type parameter is
unknown in given place. Putting NestedMember[Any] is not a solution as
it indicates knowledge about type parameter (in this case Any) which is
wrong as type parameter is really not known.

Signed-off-by: Grzegorz Kossakowski <grek@google.com>