Implemented Modification class which applies modifications to Messages.
Modification takes arbitrary number of modification command (calls to
insert, modify and remove methods) and builds internal structure that is
responsible for:
  * checking for conflicting modifications
  * scheduling order of modification appliance to be optimal one

That internal structure is a tree and modifications are applied in
depth-first order.

Tests cover more complicated scenarios like few modifications of the
same path in a tree and try to test for exceptions thrown when wrong
arguments are passed.

Signed-off-by: Grzegorz Kossakowski <grek@google.com>