Implemented modifications Gimd database stored using JGit library.
While this functionality a number of changes to API has been made, mainly:
  * functionality of DatabaseSpi trait has been moved to Database trait
  * code from JGitProvider has been moved to JGitSnapshot as all operations
    should be executed on specific Snapshot of Database

Modifications itself are implemented by taking function that takes Snapshot
and returns DatabaseModification object which carries information about
modifications that should be applied on top of passed Snapshot. Then,
modified Snapshot is tried to be safely merged with latest Snapshot of
Database.

Also, made JGitDatabaseException to store reference to JGit repository
that was in use when exception had been thrown.

Change-Id: I3f1cbcc89559569c831441e87c533627f3a78c10
Signed-off-by: Grzegorz Kossakowski <grek@google.com>