Added protected name: Message => String function to FileType.
The idea behind this change is that the name of every file in Gimd
database is a function from its content. Thus we require that FileType
provides an implementation of name function that returns name for a given
message.

Since pathPrefix is used for obtaining whole path for file it makes sense
to put additional condition on it that pathPrefix must end with "/"
character.

Apart from that pathPrefix, pathSuffix and userTypes are vals now as they
are required to be stable (always return the same value) so it's better
thant compiler enforces that.

Signed-off-by: Grzegorz Kossakowski <grek@google.com>