Add initial version of Java glue code useful when calling Gimd from Java.
This is initial version of Java glue code that might be useful when calling
Gimd API which heavily relies on passing functions that is poorly
supported in Java.

Right now it's a simple interface for java function and a few conversion
methods but this might evolve into something more sophisticated and then
better documented.

Change-Id: Ieaea6194c982cc96fed860ae2afea454637434c7
Signed-off-by: Grzegorz Kossakowski <grek@google.com>