Expose latestSnapshot in Database trait.
It turns out that if one wants to write a wrapper around Database then need
for latestSnapshot value is inevitable. Since accessing Scala objects from
Java is very painful it makes to write such wrappers.

Until we find better way to deal with this issue exposing latestSnapshot
as public method is the easiest solution.

Change-Id: I06b1e79e89864827c5ee63915e385322097e1638
Signed-off-by: Grzegorz Kossakowski <grek@google.com>