Check if Message that is being parsed has fields in appropriate order.
Added a check to a parser for order of fields. Check for order of fields
will be performed globally for one message. In principle it should be
possible to perform this check locally as one needs an access only to
successive parsed fields but I didn't know how to make it fit into
combinator parsing library constructs.

Change-Id: I605fbf39db62a5826678ed1bb19aacfdaad1677f
Signed-off-by: Grzegorz Kossakowski <grek@google.com>