Define implicit conversions from FieldSpec to functions that extract values.
Defined FieldSpecOne2T and FieldSpecOption implicit conversions where T is
a type of a value that given UserObject stores in it's class field. Thanks
to that it's possible to use FieldSpec as functions: Message => T which is very convenient
for defining UserType and implementing it's method toUserObject.

Another improvement is defining implicit conversion
FieldSpec => List[FieldSpec] that allows one to define list of FieldSpecs
in an elegant way.

FieldSpecMany and it's implicit conversions has not been implemented yet as
it requires much more work. The problem is with sorting of fields that can
be either interpreted as string and int fields. This ambiguity makes it
much more difficult to define what is a valid (sorted) representation of
given message that has fields named the same but storing both ints and
non-int values.

Change-Id: Iadeec3a97c7dac60508245e8f773e408ea9e5541
Signed-off-by: Grzegorz Kossakowski <grek@google.com>