Make imports more Scala 2.8-friendly.
In Scala 2.8 there has been a change in which scopes are introduced by
import clauses. This enforced the need of using absolute paths to imported
packages.

This change is backwards-compatible with Scala 2.7.5.

The change to Scala's scope handling is explained here:
http://stackoverflow.com/questions/1243794/what-are-the-biggest-differences-between-scala-2-8-and-scala-2-7/2550270#2550270

This change does not resolve all problems with porting Gimd's code to
Scala 2.8 but it's enough to make the latest version of IntelliJ happy.

Change-Id: Idfa91fe5d30f97c1918eecc3be2ac8fe51d09f92
Signed-off-by: Grzegorz Kossakowski <grzegorz.kossakowski@gmail.com>