Add initial code for the Query DSL.
This commit adds the code defines basic AST nodes and operations that allow
one to construct ASTs.

Both form basic, type-safe query API for Gimd. Examples of the API usage
can be found in tests that are provided.

This code adds support for FieldSpecOne variant of FieldSpec. It's left to
implement handling of optional values and lists which will support
different set of operations.

Change-Id: Ie7886ee012c4286a15289b14aa4b09be060c7bf3
Signed-off-by: Grzegorz Kossakowski <grzegorz.kossakowski@gmail.com>