Add tests for TreeMessageGenerator's statistical characteristics.
Logic inside of TreeMessageGenerator is non-trivial but breaking it is
rather easy task thus we need some basic statistical measures to ensure
that nothing is broken with it.

It would be good to study general case of generation of random, recursive
data structures so some magic parameters and measure have more foundation.

Change-Id: I5be7cac6e750f4bc7912deffb4444548a0392c14
Signed-off-by: Grzegorz Kossakowski <grzegorz.kossakowski@gmail.com>