Fix bug in Formatter where string values where not quoted sometimes.
Formatter was emitting output that should be parsed as string but were
parsed as numeric field. The only solution to this problem that I can
see is to check if given string can be interpreted as something else
than StringField and then put it's value in quotes which forces
parsing input into StringField.

This change fixes Issue 2.

Added ScalaCheck-styles properties that specify proper behaviour of
Formatter-Parser pair. These properties state that formatting and parsing
operations should be inverses to each other. Along with random generation
of inputs these tests cover complex functionality of parser and formatter
fairly well.

The change to Formatter may have negative impact on it's performance and
can make some Gimd data files look more noisy than we would like but I fail
to see any other solution at the moment.

One solution would be to provide automatic conversions after parsing phase
is finished in case that some input was interpreted in a way we don't like.
However, such automatic conversions cannot be made to be stable with regard
to ordering of field domains thus breaking invariant of Message that it's
fields are always sorted.

Change-Id: I64ac60bf655b1ab80d9ed10bdbfb7a09ad6ffcb3
Signed-off-by: Grzegorz Kossakowski <grzegorz.kossakowski@gmail.com>