Implement Predicate building out of Query class that stores AST.
Query class has been modified to store Manifest instance for a type
of objects query should return. This is needed because it should be
possible to convert Query into Predicate instance that requires
Manifest as a parameter.

This conversion is already implemented by Query.predicate which makes
use of PredicateBuilder.

Change-Id: I9b62c05946e344e19acfdc3809c51b7074c454d7
Signed-off-by: Grzegorz Kossakowski <grzegorz.kossakowski@gmail.com>