Create repository if missing and make deletion configurable.
This is a quick hack so tests can leave generated repository for later
use if generation of repository takes too long.

This is against idea that test-cases should be self-contained and
should not leave any garbage between test executions.

Also added code responsible for creation of initial commit. That
commit contains empty tree.

Change-Id: Ia26354e4226ac1e299ffd14c7897bf626b2a3d32
Signed-off-by: Grzegorz Kossakowski <grzegorz.kossakowski@gmail.com>