Optimise Parser by avoiding unnecessary allocations and computations.
Parsers numeric and timestamp do not depend on name for *parsing* but
only for producing the result of given type (Field which has a name).
Therefore it's more logical that these parsers will be of type
Parser[String => Field] which directly expresses true dependency on
name just for producing the final result but not for parsing.

This way previous def which constructed new parser for every call can
be replaced by val so the same parser is reused all the time.

This trick is applied to parsers that didn't depend on any input so
defs could be replaced by vals right away.

TextLanguageRules has been updated accordingly to this change.

Change-Id: I48c87b26f66377cdaa2d5cb541c902d0a407a3f6
Signed-off-by: Grzegorz Kossakowski <grzegorz.kossakowski@gmail.com>