Fixed two bugs in lucene.Indexer class.
First bug occurred when it Indexer tried to index an empty repository.
Lucene tracks some internal update counter that would be not touched
due to empty repository thus not committing index for empty
repository. This broke Indexer because it would leave Lucene in an
inconsistent state.

Second bug was in incremental indexing. We are asking TreeWalk for
data coming from the first tree which should be target tree and not
base tree.

Added a test-case that covers both of these problems.

Change-Id: I5f85c331ec0d44a47ffc89bd297c31333b94494a
Signed-off-by: Grzegorz Kossakowski <grzegorz.kossakowski@gmail.com>