JGitDatabase.updateRef delegates its task to an actor.
Update of reference is being delegated to special actor that tries
to update reference for given span of time and gives up if it's not
possible. The reason why it may not be possible may be problem with
obtaining a lock due to other threads updating the reference.

Added test-case that covers concurrent modifications of JGitDatabase
with and without Lucene optimization. This test creates very nice
looking history full of branches and merges that shows that JGti logic
works ok.

Change-Id: I4cb0204d7902760c63f31a0e797167430ab48438
Signed-off-by: Grzegorz Kossakowski <grzegorz.kossakowski@gmail.com>