JGit does not support criss-cross merges so Gimd has to deal with them
In case of criss-cross merges JGit complain with IOException with
message saying "Multiple merges bases for [...]". The only solution is
to rebase the change. Therefore we need to catch the exception with
ugly testing on a message and return None result in merge method that
signalises that merge is not possible to be performed and change has
to be rebased.

This problem clearly exhibits overuse of exceptions for informing
about expected result. It quite nasty problem in this case because
only concurrent modifications to Gimd db can exhibit this issue.
Actually there already exists better way of informing about the result
by using another case to JGit's Result class.

It's probably a good suggestion to JGit project.

Change-Id: I06c9c735f5a57610e72bdf517f2aec8e74db199e
Signed-off-by: Grzegorz Kossakowski <grzegorz.kossakowski@gmail.com>