Handle Result.NO_CHANDE returned by JGit's updateRef method.
In this case it's the best to quote the comment from the code:

NO_CHANGE is returned in case that both onto and commitId are the
same. From Gimd's point of view it means that some concurrent change
has been applied that brought Gimd's db to the exactly the same state
we are trying to commit here. The only way to handle this case is to
reset the transaction. We do that by returning None which means that
current modification cannot be applied and the transaction has to be
restarted.

Change-Id: Id827f24920c4a3258656f2179d20235e657b7166
Signed-off-by: Grzegorz Kossakowski <grzegorz.kossakowski@gmail.com>