Introduce caching of IndexWriter in Indexer.
It turned out that creation and closing of IndexWriter is very
expensive operation so we are using simple caching technique. Details
of caching are explained in code's comment.

This change introduces close() method that has to be called once given
Indexer instance is not needed anymore. Other code will have to be
adapted to this change.

Change-Id: I820f35db5928071ab55566c82f62d19caf4e89d3
Signed-off-by: Grzegorz Kossakowski <grzegorz.kossakowski@gmail.com>