Introduce AbstractJGitTestCase.withDb method and use it in all tests.
This method takes care of initializing and cleaning up of instances of
Gimd database. This is especially important because JGit-based
implementation of Gimd Database has a close() method that should be
called once we are done with given instance of Database.

In similar fashion we introduce LuceneTestCase trait that can be
mixed-in for testing of Lucene-optimized variant of JGit Database.

Change-Id: I7676df9322999b8a9e819fafbc46881f38ddcfc5
Signed-off-by: Grzegorz Kossakowski <grzegorz.kossakowski@gmail.com>